requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
bitarray==2.9.2
certifi==2024.8.30
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
cytoolz==1.0.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
eval_type_backport==0.2.0
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jsonpath-python==1.0.6
mistralai==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
parsimonious==0.10.0
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pyunormalize==16.0.0
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
toolz==1.0.0
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
web3==7.3.0
websockets==13.1
yarl==1.13.1
